# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: GreenRoot API
  version: 0.1.0
  description: Modular Go chi backend for the GreenRoot platform.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/plants:
    get:
      tags: [Plants]
      summary: List plants
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: integer
        - name: plant_type
          in: query
          schema:
            type: string
        - name: light_requirement
          in: query
          schema:
            type: string
        - name: water_requirement
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [id, scientific_name, common_name, plant_type, light_requirement, water_requirement, created_at, updated_at]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Plant catalog
    post:
      tags: [Plants]
      summary: Create plant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantRequest"
      responses:
        "201":
          description: Plant created
...
  /api/v1/plants/{id}:
    get:
      tags: [Plants]
      summary: Get plant
      parameters:
        - $ref: "#/components/parameters/PlantID"
      responses:
        "200":
          description: Plant detail
    put:
      tags: [Plants]
      summary: Update plant
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlantID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantRequest"
      responses:
        "200":
          description: Plant updated
    delete:
      tags: [Plants]
      summary: Delete plant
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlantID"
      responses:
        "200":
          description: Plant deleted
...
  /api/v1/nurseries:
    get:
      tags: [Nurseries]
      summary: List nurseries
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: nursery_status
          in: query
          schema:
            type: string
        - name: verification_status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Nursery list
    post:
      tags: [Nurseries]
      summary: Create nursery
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NurseryRequest"
      responses:
        "201":
          description: Nursery created
  /api/v1/nurseries/{id}:
    get:
      tags: [Nurseries]
      summary: Get nursery
      parameters:
        - $ref: "#/components/parameters/NurseryID"
      responses:
        "200":
          description: Nursery detail
    put:
      tags: [Nurseries]
      summary: Update nursery
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/NurseryID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NurseryRequest"
      responses:
        "200":
          description: Nursery updated
    delete:
      tags: [Nurseries]
      summary: Delete nursery
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/NurseryID"
      responses:
        "200":
          description: Nursery deleted
...
components:
  parameters:
...
    PlantID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    NurseryID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PlantRequest:
      type: object
      required: [scientific_name]
      properties:
        scientific_name:
          type: string
          example: Ficus religiosa
        common_name:
          type: string
          nullable: true
          example: Peepal
        english_description:
          type: string
          nullable: true
        plant_type:
          type: string
          nullable: true
          example: TREE
        light_requirement:
          type: string
          nullable: true
          example: FULL_SUN
        water_requirement:
          type: string
          nullable: true
          example: MEDIUM
        category_ids:
          type: array
          items:
            type: integer
        care_guide:
          $ref: "#/components/schemas/CareGuideRequest"
...
    CareGuideRequest:
      type: object
      properties:
        sunlight:
          type: string
          nullable: true
        watering:
          type: string
          nullable: true
        soil:
          type: string
          nullable: true
        temperature:
          type: string
          nullable: true
        fertilizer:
          type: string
          nullable: true
        pruning:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    NurseryRequest:
      type: object
      required: [name]
      properties:
        code:
          type: string
          nullable: true
        name:
          type: string
        gst_number:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
          enum: [ACTIVE, INACTIVE, SUSPENDED, DELETED]
...